move planet sprite to sprite dir
